Platforms/RPi4: Add Trusted Firmware binaries
Two versions are provided so that we can switch between using PL011 or
the 'mini UART' for serial output.

These firmwares were built using a custom version of TF-A, since the
official repository is currently missing a fix to get proper baudrate
when using a recent version of start4.elf, as we as the ability to
switch between PL011 and mini UART.

The firmwares are built for a dtb base address of 0x00020000 and a
UEFI payload base address of 0x00030000.

Signed-off-by: Pete Batard <pete@akeo.ie>